backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
flasgger==0.9.7.1
PyJWT==2.8.0
Werkzeug==3.0.0
